refactor(soul): extract FlowRunner into its own module
Move FlowRunner and the flow constants out of the 2209-line PythinkerSoul
module into soul/flow_runner.py. FlowRunner collaborates with the soul only
through its public turn machinery, so the dependency is type-only
(TYPE_CHECKING); FLOW_COMMAND_PREFIX is re-exported from pythinkersoul to
preserve the shell UI's import path. No behavior change. First step of the
staged PythinkerSoul (Phase A) decomposition.
