+++ src/pythinker_code/soul/pythinkersoul.py
@@ -41,7 +41,6 @@ from pythinker_code.notifications import (
 )
 from pythinker_code.prompt_templates import PromptTemplate, expand_prompt_template
 from pythinker_code.skill import Skill, read_skill_text_with_local_specialization
-from pythinker_code.skill.flow import Flow, FlowEdge, FlowNode, parse_choice
 from pythinker_code.soul import (
     LLMNotSet,
     LLMNotSupported,
@@ -74,6 +73,7 @@ from pythinker_code.soul.dynamic_injection import (
 )
 from pythinker_code.soul.dynamic_injections.auto_mode import AutoModeInjectionProvider
 from pythinker_code.soul.dynamic_injections.plan_mode import PlanModeInjectionProvider
+from pythinker_code.soul.flow_runner import FLOW_COMMAND_PREFIX, FlowRunner
 from pythinker_code.soul.message import (
     check_message,
     system,
@@ -125,8 +125,6 @@ if TYPE_CHECKING:
 
 
 SKILL_COMMAND_PREFIX = "skill:"
-FLOW_COMMAND_PREFIX = "flow:"
-DEFAULT_MAX_FLOW_MOVES = 1000
 
 
 def _safe_cwd(fallback: str) -> str:
@@ -2028,182 +2026,3 @@ class BackToTheFuture(Exception):
     def __init__(self, checkpoint_id: int, messages: Sequence[Message]):
         self.checkpoint_id = checkpoint_id
         self.messages = messages
-
-
-class FlowRunner:
-    def __init__(
-        self,
-        flow: Flow,
-        *,
-        name: str | None = None,
-        max_moves: int = DEFAULT_MAX_FLOW_MOVES,
-    ) -> None:
-        self._flow = flow
-        self._name = name
-        self._max_moves = max_moves
-
-    @staticmethod
-    def ralph_loop(
-        user_message: Message,
-        max_ralph_iterations: int,
-    ) -> FlowRunner:
-        prompt_content = list(user_message.content)
-        prompt_text = Message(role="user", content=prompt_content).extract_text(" ").strip()
-        total_runs = max_ralph_iterations + 1
-        if max_ralph_iterations < 0:
-            total_runs = 1000000000000000  # effectively infinite
-
-        nodes: dict[str, FlowNode] = {
-            "BEGIN": FlowNode(id="BEGIN", label="BEGIN", kind="begin"),
-            "END": FlowNode(id="END", label="END", kind="end"),
-        }
-        outgoing: dict[str, list[FlowEdge]] = {"BEGIN": [], "END": []}
-
-        nodes["R1"] = FlowNode(id="R1", label=prompt_content, kind="task")
-        nodes["R2"] = FlowNode(
-            id="R2",
-            label=(
-                f"{prompt_text}. (You are running in an automated loop where the same "
-                "prompt is fed repeatedly. Only choose STOP when the task is fully complete. "
-                "Including it will stop further iterations. If you are not 100% sure, "
-                "choose CONTINUE.)"
-            ).strip(),
-            kind="decision",
-        )
-        outgoing["R1"] = []
-        outgoing["R2"] = []
-
-        outgoing["BEGIN"].append(FlowEdge(src="BEGIN", dst="R1", label=None))
-        outgoing["R1"].append(FlowEdge(src="R1", dst="R2", label=None))
-        outgoing["R2"].append(FlowEdge(src="R2", dst="R2", label="CONTINUE"))
-        outgoing["R2"].append(FlowEdge(src="R2", dst="END", label="STOP"))
-
-        flow = Flow(nodes=nodes, outgoing=outgoing, begin_id="BEGIN", end_id="END")
-        max_moves = total_runs
-        return FlowRunner(flow, max_moves=max_moves)
-
-    async def run(self, soul: PythinkerSoul, args: str) -> None:
-        if args.strip():
-            command = f"/{FLOW_COMMAND_PREFIX}{self._name}" if self._name else "/flow"
-            logger.warning("Agent flow {command} ignores args: {args}", command=command, args=args)
-            return
-        if self._name:
-            from pythinker_code.telemetry import track
-
-            track("flow_invoked", flow_name=self._name)
-
-        current_id = self._flow.begin_id
-        moves = 0
-        total_steps = 0
-        while True:
-            node = self._flow.nodes[current_id]
-            edges = self._flow.outgoing.get(current_id, [])
-
-            if node.kind == "end":
-                logger.info("Agent flow reached END node {node_id}", node_id=current_id)
-                return
-
-            if node.kind == "begin":
-                if not edges:
-                    logger.error(
-                        'Agent flow BEGIN node "{node_id}" has no outgoing edges; stopping.',
-                        node_id=node.id,
-                    )
-                    return
-                current_id = edges[0].dst
-                continue
-
-            if moves >= self._max_moves:
-                raise MaxStepsReached(total_steps)
-            next_id, steps_used = await self._execute_flow_node(soul, node, edges)
-            total_steps += steps_used
-            if next_id is None:
-                return
-            moves += 1
-            current_id = next_id
-
-    async def _execute_flow_node(
-        self,
-        soul: PythinkerSoul,
-        node: FlowNode,
-        edges: list[FlowEdge],
-    ) -> tuple[str | None, int]:
-        if not edges:
-            logger.error(
-                'Agent flow node "{node_id}" has no outgoing edges; stopping.',
-                node_id=node.id,
-            )
-            return None, 0
-
-        base_prompt = self._build_flow_prompt(node, edges)
-        prompt = base_prompt
-        steps_used = 0
-        while True:
-            result = await self._flow_turn(soul, prompt)
-            steps_used += result.step_count
-            if result.stop_reason == "tool_rejected":
-                logger.error("Agent flow stopped after tool rejection.")
-                return None, steps_used
-
-            if node.kind != "decision":
-                return edges[0].dst, steps_used
-
-            choice = (
-                parse_choice(result.final_message.extract_text(" "))
-                if result.final_message
-                else None
-            )
-            next_id = self._match_flow_edge(edges, choice)
-            if next_id is not None:
-                return next_id, steps_used
-
-            options = ", ".join(edge.label or "" for edge in edges)
-            logger.warning(
-                "Agent flow invalid choice. Got: {choice}. Available: {options}.",
-                choice=choice or "<missing>",
-                options=options,
-            )
-            prompt = (
-                f"{base_prompt}\n\n"
-                "Your last response did not include a valid choice. "
-                "Reply with one of the choices using <choice>...</choice>."
-            )
-
-    @staticmethod
-    def _build_flow_prompt(node: FlowNode, edges: list[FlowEdge]) -> str | list[ContentPart]:
-        if node.kind != "decision":
-            return node.label
-
-        if not isinstance(node.label, str):
-            label_text = Message(role="user", content=node.label).extract_text(" ")
-        else:
-            label_text = node.label
-        choices = [edge.label for edge in edges if edge.label]
-        lines = [
-            label_text,
-            "",
-            "Available branches:",
-            *(f"- {choice}" for choice in choices),
-            "",
-            "Reply with a choice using <choice>...</choice>.",
-        ]
-        return "\n".join(lines)
-
-    @staticmethod
-    def _match_flow_edge(edges: list[FlowEdge], choice: str | None) -> str | None:
-        if not choice:
-            return None
-        for edge in edges:
-            if edge.label == choice:
-                return edge.dst
-        return None
-
-    @staticmethod
-    async def _flow_turn(
-        soul: PythinkerSoul,
-        prompt: str | list[ContentPart],
-    ) -> TurnOutcome:
-        wire_send(TurnBegin(user_input=prompt))
-        res = await soul._turn(Message(role="user", content=prompt))  # type: ignore[reportPrivateUsage]
-        wire_send(TurnEnd())
-        return res
